split up tests by module
